Add new plugin events / change plugin organization (#536)
* refactor: make file per event

* feat: edit plugins organization
