+++ pumpkin/src/plugin/api/events/player/mod.rs
@@ -1,0 +1,17 @@
+pub mod player_join;
+pub mod player_leave;
+
+use std::sync::Arc;
+
+use crate::entity::player::Player;
+
+/// A trait representing events related to players.
+///
+/// This trait provides a method to retrieve the player associated with the event.
+pub trait PlayerEvent: Send + Sync {
+    /// Retrieves a reference to the player associated with the event.
+    ///
+    /// # Returns
+    /// A reference to the `Arc<Player>` involved in the event.
+    fn get_player(&self) -> &Arc<Player>;
+}
